+++ packages/sdk/src/evm/core/sdk.ts
@@ -14,8 +14,6 @@ import { SmartContract } from "../contracts/smart-contract";
 import { AbiSchema } from "../schema";
 import { SDKOptions } from "../schema/sdk-options";
 import { CurrencyValue } from "../types/index";
-import type { AbstractWallet } from "../wallets";
-import { WalletAuthenticator } from "./auth/wallet-authenticator";
 import type { ContractMetadata } from "./classes";
 import { ContractDeployer } from "./classes/contract-deployer";
 import { ContractPublisher } from "./classes/contract-publisher";
@@ -35,6 +33,7 @@ import type {
 import { UserWallet } from "./wallet/UserWallet";
 import IThirdwebContractABI from "@thirdweb-dev/contracts-js/dist/abis/IThirdwebContract.json";
 import { ThirdwebStorage } from "@thirdweb-dev/storage";
+import type { EVMWallet } from "@thirdweb-dev/wallets";
 import { Contract, ContractInterface, ethers, Signer } from "ethers";
 import invariant from "tiny-invariant";
 
@@ -63,7 +62,7 @@ export class ThirdwebSDK extends RPCConnectionHandler {
    * @beta
    */
   static async fromWallet(
-    wallet: AbstractWallet,
+    wallet: EVMWallet,
     network: ChainOrRpc,
     options: SDKOptions = {},
     storage: ThirdwebStorage = new ThirdwebStorage(),
@@ -75,7 +74,11 @@ export class ThirdwebSDK extends RPCConnectionHandler {
       ? getReadOnlyProvider(signerOrProvider)
       : signerOrProvider;
 
-    const signer = await wallet.getSigner(provider);
+    let signer = await wallet.getSigner();
+
+    if (!!provider) {
+      signer = signer.connect(provider);
+    }
 
     return ThirdwebSDK.fromSigner(signer, network, options, storage);
   }
@@ -177,10 +180,6 @@ export class ThirdwebSDK extends RPCConnectionHandler {
    * Upload and download files from IPFS or from your own storage service
    */
   public storage: ThirdwebStorage;
-  /**
-   * Enable authentication with the connected wallet
-   */
-  public auth: WalletAuthenticator;
 
   constructor(
     network: ChainOrRpc | SignerOrProvider,
@@ -193,7 +192,6 @@ export class ThirdwebSDK extends RPCConnectionHandler {
     this.storage = storage;
     this.wallet = new UserWallet(signerOrProvider, options);
     this.deployer = new ContractDeployer(signerOrProvider, options, storage);
-    this.auth = new WalletAuthenticator(signerOrProvider, this.wallet, options);
     this.multiChainRegistry = new MultichainRegistry(
       signerOrProvider,
       this.storageHandler,
@@ -203,6 +201,15 @@ export class ThirdwebSDK extends RPCConnectionHandler {
       signerOrProvider,
       this.options,
       this.storageHandler,
+    );
+  }
+
+  get auth() {
+    throw new Error(
+      `The sdk.auth namespace has been moved to the @thirdweb-dev/auth package and is no longer available after @thirdweb-dev/sdk >= 3.7.0. 
+      Please visit https://portal.thirdweb.com/auth for instructions on how to switch to using the new auth package (@thirdweb-dev/auth@3.0.0).
+      
+      If you still want to use the old @thirdweb-dev/auth@2.0.0 package, you can downgrade the SDK to version 3.6.0.`,
     );
   }
 
@@ -584,7 +591,6 @@ export class ThirdwebSDK extends RPCConnectionHandler {
 
   private updateContractSignerOrProvider() {
     this.wallet.connect(this.getSignerOrProvider());
-    this.auth.updateSignerOrProvider(this.getSignerOrProvider());
     this.deployer.updateSignerOrProvider(this.getSignerOrProvider());
     this._publisher.updateSignerOrProvider(this.getSignerOrProvider());
     this.multiChainRegistry.updateSigner(this.getSignerOrProvider());