+++ packages/sdk/src/evm/wallets/types/abstract.ts
@@ -1,7 +1,0 @@
-import { ethers } from "ethers";
-
-export abstract class AbstractWallet {
-  abstract getSigner(
-    provider?: ethers.providers.Provider,
-  ): Promise<ethers.Signer>;
-}