+++ packages/wallets/src/evm/wallets/private-key.ts
@@ -1,0 +1,13 @@
+import { AbstractWallet } from "./abstract";
+import { ethers } from "ethers";
+
+export class PrivateKeyWallet extends AbstractWallet {
+  constructor(privateKey: string) {
+    super();
+    this.signer = new ethers.Wallet(privateKey);
+  }
+
+  async getSigner(): Promise<ethers.Signer> {
+    return this.signer as ethers.Signer;
+  }
+}